requirements.txt
paramiko~=3.0.0
pywinrm~=0.4.3
pip~=22.3.1
cryptography~=39.0.0
pywin32~=305
simplejson~=3.18.1
xmltodict~=0.13.0
requests~=2.28.2
six~=1.16.0
certifi~=2022.12.7
bcrypt~=4.0.1
idna~=3.4
urllib3~=1.26.14
pycparser~=2.21
